updates subgraph drops the Upgradeable suffix from Token contract
