+++ subgraph/subgraph.template.yaml
@@ -99,14 +99,10 @@ dataSources:
       eventHandlers:
         - event: FractionsCreated(indexed uint256,indexed uint256,indexed address,address,uint256,string,string,string)
           handler: handleFractionsCreated
-        # - event: TransferSingle(indexed address,indexed address,indexed address,uint256,uint256)
-        #   handler: handleTransferSingle
         - event: SalesActivated(uint256,address,uint256)
           handler: handleSalesActivated
         - event: TermsAccepted(indexed uint256,indexed address,bytes)
           handler: handleTermsAccepted
-        - event: SharesClaimed(indexed uint256,indexed address,uint256)
-          handler: handleSharesClaimed
       file: ./src/fractionalizerMapping.ts
 templates:
   - name: FractionalizedToken
@@ -127,3 +123,5 @@ templates:
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handleTransfer
+        - event: SharesClaimed(indexed uint256,indexed address,uint256)
+          handler: handleSharesClaimed
